requirements.txt
absl-py==2.1.0
addict==2.4.0
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
automodel==0.5.9
beautifulsoup4==4.12.2
bleach==6.1.0
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
cachetools==5.3.3
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi==1.15.1
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cifar2png==0.0.4
click==8.1.3
colorama==0.4.6
contourpy==1.1.1
cramjam==2.8.3
crcmod==1.7
cryptography==42.0.5
cycler==0.12.1
defusedxml==0.7.1
einops==0.7.0
exceptiongroup==1.1.1
fastjsonschema==2.17.1
fastparquet==2024.2.0
filelock @ file:///croot/filelock_1700591183607/work
fonttools==4.49.0
fsspec==2024.2.0
gast==0.3.3
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
google==3.0.0
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
huggingface-hub==0.21.4
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==6.6.0
importlib_resources==6.1.2
ipywidgets==8.0.6
Jinja2==3.1.2
jmespath==0.10.0
jupyter==1.0.0
jupyter-events==0.6.3
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kaggle==1.6.11
Keras==2.4.3
kiwisolver==1.4.5
Markdown==3.5.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mat4py==0.6.0
matplotlib==3.7.5
mdurl==0.1.2
mistune==2.0.5
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mmcls==0.25.0
mmcv==2.1.0
mmcv-full==1.7.2
-e git+https://github.com/open-mmlab/mmengine.git@2fe0ecec3dd2fc2be604fe6cea2558b7d7e3e041#egg=mmengine
model-index==0.1.11
modelindex==0.0.2
mpmath @ file:///croot/mpmath_1690848262763/work
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.1
onnx==1.16.1
opencv-contrib-python==4.6.0.66
opencv-python==4.7.0.72
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.34
ordered-set==4.1.0
oss2==2.17.0
overrides==7.3.1
packaging==23.2
panda==0.3.1
pandas==2.0.3
pandocfilters==1.5.0
pathlib==1.0.1
pillow @ file:///croot/pillow_1707233021655/work
platformdirs==3.5.1
plumbum==1.8.2
prometheus-client==0.17.0
proto-plus==1.23.0
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
Pygments==2.14.0
pyparsing==3.1.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2023.4
PyYAML @ file:///croot/pyyaml_1698096049011/work
qtconsole==5.4.3
QtPy==2.3.1
regex==2023.12.25
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpyc==6.0.0
rsa==4.9
safetensors==0.4.2
scipy==1.4.1
seaborn==0.12.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sympy @ file:///croot/sympy_1701397643339/work
tabulate==0.9.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
termcolor==2.4.0
terminado==0.17.1
text-unidecode==1.3
tinycss2==1.2.1
tokenizers==0.13.3
torch==1.9.1+cu111
torchaudio==0.9.1
torchvision==0.10.1+cu111
tornado==6.3.2
tqdm==4.65.2
traitlets==5.9.0
triton==2.1.0
typing_extensions==4.5.0
tzdata==2024.1
ultralytics==8.2.28
ultralytics-thop==0.2.7
uritemplate==4.1.1
urllib3==1.26.13
webencodings==0.5.1
websocket-client==1.5.2
widgetsnbextension==4.0.7
yapf==0.32.0
zipp==3.17.0
